+++ lib/kb_concoct/Utils/ConcoctUtil.py
@@ -7,8 +7,6 @@ import time
 import uuid
 import zipfile
 
-from Bio import SeqIO
-
 from installed_clients.AssemblyUtilClient import AssemblyUtil
 from installed_clients.DataFileUtilClient import DataFileUtil
 from installed_clients.KBaseReportClient import KBaseReport
@@ -16,16 +14,18 @@ from installed_clients.MetagenomeUtilsClient import MetagenomeUtils
 from installed_clients.ReadsUtilsClient import ReadsUtils
 from installed_clients.WorkspaceClient import Workspace
 
+
 def log(message, prefix_newline=False):
     """Logging function, provides a hook to suppress or redirect log messages."""
     print(('\n' if prefix_newline else '') + '{0:.2f}'.format(time.time()) + ': ' + str(message))
 
 
 class ConcoctUtil:
-    CONCOCT_TOOLKIT_PATH = '/kb/deployment/bin/CONCOCT/bin'
-    CONCOCT_BIN_DIR = 'final_bins'
-    MAPPING_THREADS=16
-    BBMAP_MEM='30g'
+    CONCOCT_BASE_PATH = '/kb/deployment/bin/CONCOCT'
+    CONCOCT_RESULT_DIRECTORY = 'concoct_output_dir'
+    CONCOCT_BIN_RESULT_DIR = 'final_bins'
+    MAPPING_THREADS = 16
+    BBMAP_MEM = '30g'
 
     def __init__(self, config):
         self.callback_url = config['SDK_CALLBACK_URL']
@@ -38,23 +38,24 @@ class ConcoctUtil:
         self.mgu = MetagenomeUtils(self.callback_url)
 
     # Function to return object ID based on object name
-    def get_obj_id(obj_name):
+    def get_obj_id(object_name):
         """
-        Example: get_obj_id("MAG-QC_Archaea.SAGs.Prokka")
+        Example: get_obj_id("MAG-QC_Archaea.SAGs.Prokka"), returns '32342/1/3'
         """
-
-        ws = Workspace('https://appdev.kbase.us/services/ws')
+        ws = Workspace(self.ws_url)
         ws_name = os.environ['KB_WORKSPACE_ID']
         try:
-            obj_id = ws.get_object_info3({'objects': [{'workspace': ws_name, 'name': obj_name}]})['paths'][0][0]
-            return obj_id
+            object_id = ws.get_object_info3({'objects': [{'workspace': ws_name, 'name': object_name}]})['paths'][0][0]
+            return object_id
         except:
             return False
 
     # Function to return object data based on object ID
     def get_object_data(object_id):
         """
-        Fetch data from the workspace. Example1: get_object_data(get_obj_id("MAG-QC_Archaea.SAGs.RAST")); Example2: get_object_data(u'43402/2132/1')
+        Fetch data from the workspace.
+        Example1: get_object_data(get_obj_id("MAG-QC_Archaea.SAGs.RAST"))
+        Example2: get_object_data(u'43402/2132/1')
         """
         from biokbase.workspace.client import Workspace
         import os
@@ -62,26 +63,26 @@ class ConcoctUtil:
         ws = Workspace(self.ws_url)
         return ws.get_objects([{"ref":object_id}])[0]
 
-    def get_object_type(object_id):
-        """
-        Fetch data from the workspace. Example1: get_object_data(get_obj_id("MAG-QC_Archaea.SAGs.RAST")); Example2: get_object_data(u'43402/2132/1')
-        """
 
+    # Function to return object type based on object ID
+    def get_object_type(object_name):
+        """
+        Fetch data type from the workspace, for example "PairedEndLibrary".
+        Example1: get_object_type('lib2.oldstyle.fastq_reads')
+        """
         ws = Workspace(self.ws_url)
-        return ws.get_objects([{"ref":object_id, "no-data": 1}])[0]['info'][2]
+        return get_object_data(get_object_id(object_name)).values()[0][2]
 
-
-    def validate_run_concoct_params(self, params):
+    def validate_run_concoct_params(self, task_params):
         """
         validate_run_concoct_params:
                 validates params passed to run_concoct method
-
         """
         log('Start validating run_concoct params')
 
         # check for required parameters
         for p in ['assembly_ref', 'binned_contig_name', 'workspace_name', 'reads_list', 'read_mapping_tool']:
-            if p not in params:
+            if p not in task_params:
                 raise ValueError('"{}" parameter is required, but missing'.format(p))
 
     def mkdir_p(self, path):
@@ -106,7 +107,7 @@ class ConcoctUtil:
         log('Start executing command:\n{}'.format(command))
         log('Command is running from:\n{}'.format(self.scratch))
         pipe = subprocess.Popen(command, stdout=subprocess.PIPE, shell=True)
-        output,stderr = pipe.communicate()
+        output, stderr = pipe.communicate()
         exitCode = pipe.returncode
 
         if (exitCode == 0):
@@ -117,7 +118,7 @@ class ConcoctUtil:
             error_msg += 'Exit Code: {}\nOutput:\n{}\nStderr:\n{}'.format(exitCode, output, stderr)
             raise ValueError(error_msg)
             sys.exit(1)
-        return (output,stderr)
+        return (output, stderr)
 
     def stage_reads_list_file(self, reads_list):
         """
@@ -143,12 +144,10 @@ class ConcoctUtil:
 
         return result_file_path
 
-
     def get_contig_file(self, assembly_ref):
         """
-        get_contig_file: get contif file from GenomeAssembly object
+        get_contig_file: get contig file from GenomeAssembly object
         """
-
         contig_file = self.au.get_assembly_as_fasta({'ref': assembly_ref}).get('path')
 
         sys.stdout.flush()
@@ -158,7 +157,7 @@ class ConcoctUtil:
 
     def retrieve_and_clean_assembly(self, task_params):
         if os.path.exists(task_params['contig_file_path']):
-            assembly =  task_params['contig_file_path']
+            assembly = task_params['contig_file_path']
             print("FOUND ASSEMBLY ON LOCAL SCRATCH")
         else:
             # we are on njsw so lets copy it over to scratch
@@ -167,68 +166,69 @@ class ConcoctUtil:
         # remove spaces from fasta headers because that breaks bedtools
         assembly_clean = os.path.abspath(assembly).split('.fa')[0] + "_clean.fa"
 
-        command = '/bin/bash reformat.sh in={} out={} addunderscore'.format(assembly,assembly_clean)
+        command = '/bin/bash reformat.sh in={} out={} addunderscore'.format(assembly, assembly_clean)
 
         log('running reformat command: {}'.format(command))
-        out,err = self.run_command(command)
+        out, err = self.run_command(command)
 
         return assembly_clean
 
-
-    def generate_alignment_bams(self, task_params):
+    def generate_stats_for_genome_bins(self, task_params, genome_bin_fna_file, bbstats_output_file):
         """
-            This function runs the bbmap.sh alignments and creates the
-            bam files from sam files using samtools.
+        generate_command: bbtools stats.sh command
         """
-        # even though read_name is just one file, make read_name a
-        # list because that is what stage_reads_list_file expects
+        log("\n\nRunning generate_stats_for_genome_bins on {}".format(genome_bin_fna_file))
+        result_directory = os.path.join(self.scratch, self.CONCOCT_RESULT_DIRECTORY)
+        genome_bin_fna_file = os.path.join(self.scratch, self.CONCOCT_RESULT_DIRECTORY, genome_bin_fna_file)
+        command = '/bin/bash stats.sh in={} format=3 > {}'.format(genome_bin_fna_file, bbstats_output_file)
+        self.run_command(command)
+        bbstats_output = open(bbstats_output_file, 'r').readlines()[1]
+        print('bbstats_output {}'.format(bbstats_output))
+        n_scaffolds = bbstats_output.split('\t')[0]
+        n_contigs = bbstats_output.split('\t')[1]
+        scaf_bp = bbstats_output.split('\t')[2]
+        contig_bp = bbstats_output.split('\t')[3]
+        gap_pct = bbstats_output.split('\t')[4]
+        scaf_N50 = bbstats_output.split('\t')[5]
+        scaf_L50 = bbstats_output.split('\t')[6]
+        ctg_N50 = bbstats_output.split('\t')[7]
+        ctg_L50 = bbstats_output.split('\t')[8]
+        scaf_N90 = bbstats_output.split('\t')[9]
+        scaf_L90 = bbstats_output.split('\t')[10]
+        ctg_N90 = bbstats_output.split('\t')[11]
+        ctg_L90 = bbstats_output.split('\t')[12]
+        scaf_max = bbstats_output.split('\t')[13]
+        ctg_max = bbstats_output.split('\t')[14]
+        scaf_n_gt50K = bbstats_output.split('\t')[15]
+        scaf_pct_gt50K = bbstats_output.split('\t')[16]
+        gc_avg = bbstats_output.split('\t')[17]
+        gc_std = bbstats_output.split('\t')[18]
 
-        reads_list = task_params['reads_list']
+        log('Generated generate_stats_for_genome_bins command: {}'.format(command))
 
-        # we need to copy the fastq file from workspace to local scratch.
-        # 'fastq' should be path to one file (if we had read.fwd.fq and read.rev.fq then they
-        # should have been interleaved already). However, I should interleave them in case
-        # this function gets called other than the narrative.
-        # this should return a list of 1 fastq file path on scratch
-        read_scratch_path = self.stage_reads_list_file(reads_list)
+        return {'n_scaffolds': n_scaffolds, 'n_contigs': n_contigs, 'scaf_bp': scaf_bp, 'contig_bp': contig_bp, 'gap_pct': gap_pct, 'scaf_N50': scaf_N50, 'scaf_L50': scaf_L50, 'ctg_N50': ctg_N50, 'scaf_N90': scaf_N90, 'ctg_N90': ctg_N90, 'ctg_L90': ctg_L90, 'scaf_max': scaf_max, 'ctg_max': ctg_max, 'scaf_n_gt50K': scaf_n_gt50K, 'gc_avg': gc_avg, 'gc_std':gc_std}
+        # return {'i': i, 'card': card, 'other_field': other_field, ...}
 
 
-        task_params['read_scratch_file'] = read_scratch_path
 
-
-        assembly_clean = self.retrieve_and_clean_assembly(task_params)
-
-        fastq = read_scratch_path[0]
-        #reads_type = self.get_obj_id(str(os.path.basename(fastq)[1]))
-        print(os.path.basename(fastq))
-        #reads_type = self.get_object_data(self.get_obj_id(os.path.basename(fastq)[1])).values()[0][2]
-        #print("read_type is " + reads_type)
-
-        result_directory = task_params['result_directory']
-        sam = os.path.basename(fastq) + '.sam'
-        sam = os.path.join(result_directory, sam)
-        sorted_bam = sam + '.sorted.bam'
-        print("task_params are: " + str(task_params['read_mapping_tool']))
-
+    def run_read_mapping_unpaired_mode(self, task_params, assembly_clean, fastq, sam):
         if task_params['read_mapping_tool'] == 'bbmap':
-            command = '/bin/bash bbmap.sh -Xmx{} fast threads={} ref={} in={} out={} mappedonly nodisk overwrite'.format(self.BBMAP_MEM,self.MAPPING_THREADS,assembly_clean,fastq,sam)
+            command = '/bin/bash bbmap.sh -Xmx{} fast threads={} ref={} in={} out={} mappedonly nodisk overwrite'.format(self.BBMAP_MEM, self.MAPPING_THREADS, assembly_clean, fastq, sam)
         elif task_params['read_mapping_tool'] == 'bwa':
-            command = 'bwa index {} && bwa mem -t {} {} {} > {}'.format(assembly_clean,self.MAPPING_THREADS,assembly_clean,fastq,sam)
+            command = 'bwa index {} && bwa mem -t {} {} {} > {}'.format(assembly_clean, self.MAPPING_THREADS, assembly_clean, fastq, sam)
         elif task_params['read_mapping_tool'] == 'bowtie2':
             bt2index = os.path.basename(assembly_clean) + '.bt2'
-            command = 'bowtie2-build -f {} --threads {} {} && bowtie2 -x {} -U {} --threads {} -S {}'.format(assembly_clean,self.MAPPING_THREADS,bt2index,bt2index,fastq,self.MAPPING_THREADS,sam)
+            command = 'bowtie2-build -f {} --threads {} {} && bowtie2 -x {} -U {} --threads {} -S {}'.format(assembly_clean, self.MAPPING_THREADS, bt2index, bt2index, fastq, self.MAPPING_THREADS, sam)
         elif task_params['read_mapping_tool'] == 'minimap2':
-            command = 'minimap2 -ax sr -t {} {} {} > {}'.format(self.MAPPING_THREADS,assembly_clean,fastq,sam)
-
+            command = 'minimap2 -ax sr -t {} {} {} > {}'.format(self.MAPPING_THREADS, assembly_clean, fastq, sam)
         log('running alignment command: {}'.format(command))
-        out,err = self.run_command(command)
+        out, err = self.run_command(command)
 
-        task_params['sorted_bam'] = sorted_bam
-
+    def convert_sam_to_sorted_and_indexed_bam(self, sam, sorted_bam):
         # create bam files from sam files
-        command = 'samtools view -F 0x04 -uS {} | samtools sort - -o {}'.format(sam,sorted_bam)
+        command = 'samtools view -F 0x04 -uS {} | samtools sort - -o {}'.format(sam, sorted_bam)
 
-        log('running samtools command: {}'.format(command))
+        log('running samtools command to generate sorted bam: {}'.format(command))
         self.run_command(command)
 
         # verify we got bams
@@ -242,8 +242,60 @@ class ConcoctUtil:
         # index the bam file
         command = 'samtools index {}'.format(sorted_bam)
 
-        log('running samtools command: {}'.format(command))
+        log('running samtools command to index sorted bam: {}'.format(command))
         self.run_command(command)
+
+    def generate_alignment_bams(self, task_params):
+        """
+            This function runs the bbmap.sh alignments and creates the
+            bam files from sam files using samtools.
+        """
+        # first, clean the assembly file so that there are no spaces in the fasta headers
+        assembly_clean = self.retrieve_and_clean_assembly(task_params)
+
+        # even though read_name is just one file, make read_name a
+        # list because that is what stage_reads_list_file expects
+        reads_list = task_params['reads_list']
+
+        # we need to copy the fastq file from workspace to local scratch.
+        # 'fastq' should be path to one file (if we had read.fwd.fq and read.rev.fq then they
+        # should have been interleaved already). However, I should interleave them in case
+        # this function gets called other than the narrative.
+        # this should return a list of 1 fastq file path on scratch
+        read_scratch_path = self.stage_reads_list_file(reads_list)
+
+        print("read_scratch_path {}".format(read_scratch_path))
+        print("read_scratch_path {}".format(read_scratch_path[0]))
+
+        # list of reads files, can be 1 or more. assuming reads are either type unpaired or interleaved
+        # will not handle unpaired forward and reverse reads input as seperate (non-interleaved) files
+        task_params['read_scratch_file'] = read_scratch_path
+
+        for i in range(len(read_scratch_path)):
+            fastq = read_scratch_path[i]
+            #fastq = read_scratch_path[0] #needs to be a loop
+            #reads_type = self.get_obj_id(str(os.path.basename(fastq)[1]))
+            print("OS.PATH.BASENAME0 {}".format([os.path.basename(fastq)]))
+            print("OS.PATH.BASENAME1 {}".format(type([os.path.basename(fastq)])))
+
+            #print("OS.PATH.BASENAME" + str(self.get_obj_id(os.path.basename(fastq))))
+            # reads_type = self.get_object_type(str(os.path.basename(fastq)))
+            # reads_type = self.get_object_type(self.get_obj_id([os.path.basename(fastq)])).values()[0][2]
+            # print("read_type is " + reads_type)
+
+            #result_directory = task_params['result_directory']
+            sam = os.path.basename(fastq).split('.fastq')[0] + ".sam"
+            sam = os.path.join(self.CONCOCT_RESULT_DIRECTORY, sam)
+
+            self.run_read_mapping_unpaired_mode(task_params, assembly_clean, fastq, sam)
+
+            sorted_bam = os.path.abspath(sam).split('.sam')[0] + "_sorted.bam"
+
+            # need to get this to work on list
+            task_params['sorted_bam'] = sorted_bam
+
+            self.convert_sam_to_sorted_and_indexed_bam(sam, sorted_bam)
+
         return assembly_clean
 
     def generate_make_coverage_table_command(self, task_params):
@@ -252,7 +304,7 @@ class ConcoctUtil:
 
         min_contig_length = task_params['min_contig_length']
 
-        depth_file_path  = os.path.join(self.scratch, str('concoct_depth.txt'))
+        depth_file_path = os.path.join(self.scratch, str('concoct_depth.txt'))
         command = '/kb/module/lib/kb_concoct/bin/jgi_summarize_bam_contig_depths '
         command += '--outputDepth {} --minContigLength {} --minContigDepth 1 '.format(depth_file_path, min_contig_length)
         command += task_params['sorted_bam']
@@ -262,61 +314,133 @@ class ConcoctUtil:
 
         return depth_file_path
 
-    def generate_concoct_command(self, params):
-        """
-        generate_command: generate concoct params
-        """
-
+    # untested, need to check in UI
+    def fix_generate_concoct_command_ui_bug(self, task_params):
         # needed to get checkbox for UI to work with string objects, for some reason strings are converted to numerics when running inside KBase UI.
-        parameter_no_total_coverage = params.get('no_total_coverage')
-        parameter_no_cov_normalization = params.get('no_cov_normalization')
+        parameter_no_total_coverage = task_params['no_total_coverage']
+        parameter_no_cov_normalization = task_params['no_cov_normalization']
 
-        if params.get('no_total_coverage') is 1:
+        if task_params['no_total_coverage'] is 1:
             parameter_no_total_coverage = '--no_total_coverage'
-        elif params.get('no_total_coverage') is 0:
+        elif task_params['no_total_coverage'] is 0:
             parameter_no_total_coverage = ' '
 
-        if params.get('no_cov_normalization') is 1:
+        if task_params['no_cov_normalization'] is 1:
             parameter_no_cov_normalization = '--no_cov_normalization'
-        elif params.get('no_cov_normalization') is 0:
+        elif task_params['no_cov_normalization'] is 0:
             parameter_no_cov_normalization = ' '
 
-
-        print("\n\nRunning generate_concoct_command")
-
-        print("no_cov_normalization is: " + str(params.get('no_cov_normalization')))
-
-        command = 'python /kb/deployment/bin/CONCOCT/scripts/cut_up_fasta.py {} -c {} -o {} --merge_last -b temp.bed > concoct_output_dir/split_contigs.fa'.format(params.get('contig_file_path'),params.get('contig_split_size'),params.get('contig_split_overlap'))
-        command += ' && '
-        command += 'python /kb/deployment/bin/CONCOCT/scripts/concoct_coverage_table.py temp.bed concoct_output_dir/*.sorted.bam > concoct_output_dir/coverage_table.tsv'
-        command += ' && '
-        command += 'python /kb/deployment/bin/CONCOCT/bin/concoct --composition_file concoct_output_dir/split_contigs.fa -l {} -b concoct_output_dir --coverage_file concoct_output_dir/coverage_table.tsv -t {} -k {} -c {} -i {} --total_percentage_pca {} {} {}'.format(params.get('min_contig_length'),
-                                         self.MAPPING_THREADS,params.get('kmer_size'),params.get('max_clusters_for_vgmm'),params.get('max_iterations_for_vgmm'),params.get('total_percentage_pca'),parameter_no_cov_normalization,parameter_no_total_coverage)
-        command += ' && '
-        command += 'python /kb/deployment/bin/CONCOCT/scripts/merge_cutup_clustering.py concoct_output_dir/clustering_gt{}.csv > concoct_output_dir/clustering_merged.csv'.format(params.get('min_contig_length'))
-        command += ' && '
-        command += 'mkdir concoct_output_dir/{}'.format(self.CONCOCT_BIN_DIR)
-        command += ' && '
-        command += 'python /kb/deployment/bin/CONCOCT/scripts/extract_fasta_bins.py {} concoct_output_dir/clustering_merged.csv --output_path concoct_output_dir/{}'.format(params.get('contig_file_path'),self.CONCOCT_BIN_DIR)
+        return (parameter_no_total_coverage, parameter_no_cov_normalization)
 
 
-        log('Generated run_concoct command: {}'.format(command))
+
+    def generate_concoct_cut_up_fasta_command(self, task_params):
+        """
+        generate_command: concoct cut_up_fasta
+        """
+        log("\n\nRunning generate_concoct_cut_up_fasta_command")
+
+        command = 'python {}/scripts/cut_up_fasta.py {} -c {} -o {} --merge_last -b temp.bed > {}/split_contigs.fa'.format(self.CONCOCT_BASE_PATH, task_params['contig_file_path'], task_params['contig_split_size'],task_params['contig_split_overlap'],self.CONCOCT_RESULT_DIRECTORY)
+        log('Generated concoct_cut_up_fasta command: {}'.format(command))
 
         return command
 
 
-    def generate_summary_utils_command(self, params):
+
+    def generate_concoct_coverage_table_from_bam(self, task_params):
+        """
+        generate_command: concoct generate coverage table
+        """
+        log("\n\nRunning generate_concoct_coverage_table_from_bam")
+        command = 'python {}/scripts/concoct_coverage_table.py temp.bed {}/*_sorted.bam > {}/coverage_table.tsv'.format(self.CONCOCT_BASE_PATH, self.CONCOCT_RESULT_DIRECTORY, self.CONCOCT_RESULT_DIRECTORY)
+        log('Generated concoct generate coverage table from bam command: {}'.format(command))
+
+        return command
+
+
+    def generate_concoct_command(self, task_params):
+        """
+        generate_command: concoct
+        """
+        parameter_no_total_coverage, parameter_no_cov_normalization = self.fix_generate_concoct_command_ui_bug(task_params)
+
+        log("\n\nRunning generate_concoct_command")
+        command = 'python {}/bin/concoct --composition_file {}/split_contigs.fa -l {} -b {} --coverage_file {}/coverage_table.tsv -t {} -k {} -c {} -i {} --total_percentage_pca {} {} {}'.format(self.CONCOCT_BASE_PATH, self.CONCOCT_RESULT_DIRECTORY, task_params['min_contig_length'], self.CONCOCT_RESULT_DIRECTORY, self.CONCOCT_RESULT_DIRECTORY,
+                                         self.MAPPING_THREADS,task_params['kmer_size'],task_params['max_clusters_for_vgmm'],task_params['max_iterations_for_vgmm'],task_params['total_percentage_pca'],parameter_no_cov_normalization,parameter_no_total_coverage)
+        log('Generated concoct command: {}'.format(command))
+
+        return command
+
+    def generate_concoct_post_clustering_merging_command(self, task_params):
+        """
+        generate_command: concoct post cluster merging
+        """
+
+        log("\n\nRunning generate_concoct_post_clustering_merging_command")
+
+        command = 'python {}/scripts/merge_cutup_clustering.py {}/clustering_gt{}.csv > {}/clustering_merged.csv'.format(self.CONCOCT_BASE_PATH, self.CONCOCT_RESULT_DIRECTORY, task_params['min_contig_length'], self.CONCOCT_RESULT_DIRECTORY)
+        log('Generated generate_concoct_post_clustering_merging command: {}'.format(command))
+
+        return command
+
+    def generate_concoct_extract_fasta_bins_command(self, task_params):
+        """
+        generate_command: concoct extract_fasta_bins
+        """
+        log("\n\nRunning generate_concoct_extract_fasta_bins_command")
+
+        command = 'mkdir {}/{}'.format(self.CONCOCT_RESULT_DIRECTORY, self.CONCOCT_BIN_RESULT_DIR)
+        command += ' && '
+        command += 'python {}/scripts/extract_fasta_bins.py {} {}/clustering_merged.csv --output_path {}/{}'.format(self.CONCOCT_BASE_PATH, task_params['contig_file_path'],self.CONCOCT_RESULT_DIRECTORY, self.CONCOCT_RESULT_DIRECTORY, self.CONCOCT_BIN_RESULT_DIR)
+        log('Generated generate_concoct_extract_fasta_bins_command command: {}'.format(command))
+
+        return command
+
+    def generate_summary_utils_command(self, task_params):
         """
         generate_command: generate summary utils command
         """
-
-        print("\n\nRunning generate_summary_utils_command")
-
-        command = "/bin/bash /kb/module/lib/kb_concoct/bin/summary_utils.sh"
-
+        log("\n\nRunning generate_summary_utils_command")
+        if task_params['write_bins_to_fasta_files'] == 1:
+            command = "/bin/bash /kb/module/lib/kb_concoct/bin/summary_utils.sh 1"
+        else:
+            command = "/bin/bash /kb/module/lib/kb_concoct/bin/summary_utils.sh 0"
         log('Generated summary_utils command: {}'.format(command))
 
         return command
+
+
+    def make_binned_contig_summary_file_for_binning_apps(self, task_params):
+        """
+        generate_command: generate binned contig summary command
+        """
+        log("\n\nRunning make_binned_contig_summary_file_for_binning_apps")
+        path_to_concoct_result = os.path.abspath(self.CONCOCT_RESULT_DIRECTORY)
+        path_to_concoct_result_bins = '{}/{}'.format(path_to_concoct_result, self.CONCOCT_BIN_RESULT_DIR)
+        path_to_concoct_result_bins = path_to_concoct_result_bins + '/' # need to fix
+        path_to_summary_file = '{}notreal.summary'.format(path_to_concoct_result_bins) # need to fix
+        with open(path_to_summary_file, 'w+') as f:
+            f.write("Bin name\tCompleteness\tGenome size\tGC content\n")
+            for dirname, subdirs, files in os.walk(path_to_concoct_result_bins):
+                for file in files:
+                    if file.endswith('.fa'):
+                         genome_bin_fna_file = os.path.join(self.CONCOCT_BIN_RESULT_DIR, file)
+                         bbstats_output_file = os.path.join(self.scratch, self.CONCOCT_RESULT_DIRECTORY, genome_bin_fna_file).split('.fa')[0] + ".bbstatsout"
+                         bbstats_output = self.generate_stats_for_genome_bins(task_params, genome_bin_fna_file, bbstats_output_file)
+                         #f.write('{}\t{}\t{}\t0\n'.format(genome_bin_fna_file.split("/")[-1], bbstats_output['gc_avg'],bbstats_output['contig_bp']))
+        f.close()
+        log('Finished make_binned_contig_summary_file_for_binning_apps function')
+
+
+    def rename_bins_with_three_digit_number(self):
+        """
+        generate_command: generate rename bins command
+        """
+        log("\n\nRunning rename_bins_with_three_digit_number")
+
+
+        log('Finished rename_bins_with_three_digit_number function')
+
 
 
     def generate_output_file_list(self, result_directory):
@@ -329,7 +453,6 @@ class ConcoctUtil:
         output_directory = os.path.join(self.scratch, str(uuid.uuid4()))
         self.mkdir_p(output_directory)
         result_file = os.path.join(output_directory, 'concoct_result.zip')
-        report_file = None
 
         with zipfile.ZipFile(result_file, 'w',
                              zipfile.ZIP_DEFLATED,
@@ -339,15 +462,14 @@ class ConcoctUtil:
                 for file in files:
                     if (file.endswith('.sam') or file.endswith('.bam') or file.endswith('.bai') or file.endswith('.summary')):
                             continue
-                    if (dirname.endswith(self.CONCOCT_BIN_DIR)):
+                    if (dirname.endswith(self.CONCOCT_BIN_RESULT_DIR)):
                             continue
-                    zip_file.write(os.path.join(dirname, file),file)
-                if (dirname.endswith(self.CONCOCT_BIN_DIR)):
-                    baseDir=os.path.basename(dirname)
+                    zip_file.write(os.path.join(dirname, file), file)
+                if (dirname.endswith(self.CONCOCT_BIN_RESULT_DIR)):
+                    baseDir = os.path.basename(dirname)
                     for file in files:
-                        full=os.path.join(dirname, file)
-                        zip_file.write(full,os.path.join(baseDir,file))
-
+                        full = os.path.join(dirname, file)
+                        zip_file.write(full, os.path.join(baseDir, file))
 
         output_files.append({'path': result_file,
                              'name': os.path.basename(result_file),
@@ -371,10 +493,7 @@ class ConcoctUtil:
         # get summary data from existing assembly object and bins_objects
         Summary_Table_Content = ''
         Overview_Content = ''
-        (binned_contig_count, input_contig_count,
-         total_bins_count) = self.generate_overview_info(assembly_ref,
-                                                          binned_contig_obj_ref,
-                                                          result_directory)
+        (binned_contig_count, input_contig_count, total_bins_count) = self.generate_overview_info(assembly_ref, binned_contig_obj_ref, result_directory)
 
         Overview_Content += '<p>Binned contigs: {}</p>'.format(binned_contig_count)
         Overview_Content += '<p>Input contigs: {}</p>'.format(input_contig_count)
@@ -396,7 +515,6 @@ class ConcoctUtil:
                             'description': 'HTML summary report for kb_concoct App'})
         return html_report
 
-
     def generate_overview_info(self, assembly_ref, binned_contig_obj_ref, result_directory):
         """
         _generate_overview_info: generate overview information from assembly and binnedcontig
@@ -407,7 +525,6 @@ class ConcoctUtil:
         binned_contig = self.dfu.get_objects({'object_refs': [binned_contig_obj_ref]})['data'][0]
 
         input_contig_count = assembly.get('data').get('num_contigs')
-        bins_directory = os.path.join(self.scratch, result_directory, self.CONCOCT_BIN_DIR)
         binned_contig_count = 0
         total_bins_count = 0
         total_bins = binned_contig.get('data').get('bins')
@@ -417,27 +534,23 @@ class ConcoctUtil:
 
         return (binned_contig_count, input_contig_count, total_bins_count)
 
-
-    def generate_report(self, binned_contig_obj_ref, params):
+    def generate_report(self, binned_contig_obj_ref, task_params):
         """
         generate_report: generate summary report
-
         """
         log('Generating report')
 
         result_directory = os.path.join(self.scratch, "concoct_output_dir")
 
-        params['result_directory'] = result_directory
+        task_params['result_directory'] = result_directory
 
-        output_files = self.generate_output_file_list(params['result_directory'])
+        output_files = self.generate_output_file_list(task_params['result_directory'])
 
-        output_html_files = self.generate_html_report(params['result_directory'],
-                                                       params['assembly_ref'],
-                                                       binned_contig_obj_ref)
+        output_html_files = self.generate_html_report(task_params['result_directory'], task_params['assembly_ref'], binned_contig_obj_ref)
 
         report_params = {
               'message': '',
-              'workspace_name': params.get('workspace_name'),
+              'workspace_name': task_params['workspace_name'],
               'file_links': output_files,
               'html_links': output_html_files,
               'direct_html_link_index': 0,
@@ -451,15 +564,13 @@ class ConcoctUtil:
 
         return report_output
 
-
-    def createDictFromDepthfile(self, depth_file_path):
+    def create_dict_from_depth_file(self, depth_file_path):
         # keep contig order (required by metabat2)
         depth_file_dict = {}
-        with open(depth_file_path,'r') as f:
+        with open(depth_file_path, 'r') as f:
             header = f.readline().rstrip().split("\t")
             # print('HEADER1 {}'.format(header))
             # map(str.strip, header)
-
             for line in f:
                 # deal with cases were fastq name has spaces.Assume first
                 # non white space word is unique and use this as ID.
@@ -469,15 +580,11 @@ class ConcoctUtil:
                     ID = vals[0].split()[0]
                 else:
                     ID = vals[0]
-
                 depth_file_dict[ID] = vals[1:]
-
             depth_file_dict['header'] = header
-
         return depth_file_dict
 
-
-    def run_concoct(self, params):
+    def run_concoct(self, task_params):
         """
         run_concoct: concoct app
 
@@ -494,67 +601,92 @@ class ConcoctUtil:
             ref: https://github.com/BinPro/CONCOCT/blob/develop/README.md
         """
         log('--->\nrunning ConcoctUtil.run_concoct\n' +
-            'params:\n{}'.format(json.dumps(params, indent=1)))
+            'task_params:\n{}'.format(json.dumps(task_params, indent=1)))
 
-        self.validate_run_concoct_params(params)
+        self.validate_run_concoct_params(task_params)
 
-        #params['split_contig_length'] = params.get('split_contig_length')
+        contig_file = self.get_contig_file(task_params['assembly_ref'])
+        task_params['contig_file_path'] = contig_file
 
-        contig_file = self.get_contig_file(params.get('assembly_ref'))
-        params['contig_file_path'] = contig_file
+        reads_list_file = self.stage_reads_list_file(task_params['reads_list'])
+        task_params['reads_list_file'] = reads_list_file
 
-        reads_list_file = self.stage_reads_list_file(params.get('reads_list'))
-        params['reads_list_file'] = reads_list_file
-
-        result_directory = os.path.join(self.scratch, "concoct_output_dir")
-
-        params['result_directory'] = result_directory
+        result_directory = os.path.join(self.scratch, self.CONCOCT_RESULT_DIRECTORY)
 
         self.mkdir_p(result_directory)
 
         cwd = os.getcwd()
         log('changing working dir to {}'.format(result_directory))
         os.chdir(result_directory)
+
         #run alignments, and update input contigs to use the clean file
 
-        params['contig_file_path'] = self.generate_alignment_bams(params)
+        # this is the clean assembly
+        task_params['contig_file_path'] = self.generate_alignment_bams(task_params)
 
-        depth_file_path = self.generate_make_coverage_table_command(params)
+        # not used right now
+        depth_file_path = self.generate_make_coverage_table_command(task_params)
+        depth_dict = self.create_dict_from_depth_file(depth_file_path)
 
-        depth_dict = self.createDictFromDepthfile(depth_file_path)
-
-
-        #run concoct
-        command = self.generate_concoct_command(params)
+        # run concoct prep, cut up fasta input
+        command = self.generate_concoct_cut_up_fasta_command(task_params)
         self.run_command(command)
 
-        #run summary utilities
-        command = self.generate_summary_utils_command(params)
+        # run concoct make coverage table from bam
+        command = self.generate_concoct_coverage_table_from_bam(task_params)
         self.run_command(command)
 
-        #file handling and management
+        # run concoct prep and concoct
+        command = self.generate_concoct_command(task_params)
+        self.run_command(command)
+
+        # run concoct post cluster merging command
+        command = self.generate_concoct_post_clustering_merging_command(task_params)
+        self.run_command(command)
+
+        if task_params['write_bins_to_fasta_files'] == 1:
+            # run extract bins command
+            command = self.generate_concoct_extract_fasta_bins_command(task_params)
+            self.run_command(command)
+        else:
+            log('Skipping the creation of bin fasta files')
+
+        self.make_binned_contig_summary_file_for_binning_apps(task_params)
+
+        # run summary utilities and fasta renaming
+        command = self.generate_summary_utils_command(task_params)
+        self.run_command(command)
+
+
+
+        #
+        # # run summary utilities and fasta renaming
+        # command = self.generate_summary_utils_command(task_params)
+        # self.run_command(command)
+
+
+        # file handling and management
         os.chdir(cwd)
         log('changing working dir to {}'.format(cwd))
 
         log('Saved result files to: {}'.format(result_directory))
         log('Generated files:\n{}'.format('\n'.join(os.listdir(result_directory))))
 
+        # make new BinnedContig object and upload to KBase
         generate_binned_contig_param = {
-            'file_directory': os.path.join(result_directory,self.CONCOCT_BIN_DIR),
-            'assembly_ref': params['assembly_ref'],
-            'binned_contig_name': params['binned_contig_name'],
-            'workspace_name': params['workspace_name']
+            'file_directory': os.path.join(result_directory, self.CONCOCT_BIN_RESULT_DIR),
+            'assembly_ref': task_params['assembly_ref'],
+            'binned_contig_name': task_params['binned_contig_name'],
+            'workspace_name': task_params['workspace_name']
         }
-
         binned_contig_obj_ref = self.mgu.file_to_binned_contigs(generate_binned_contig_param).get('binned_contig_obj_ref')
 
-        reportVal = self.generate_report(binned_contig_obj_ref, params)
-
+        # generate report
+        reportVal = self.generate_report(binned_contig_obj_ref, task_params)
         returnVal = {
             'result_directory': result_directory,
             'binned_contig_obj_ref': binned_contig_obj_ref
         }
-
         returnVal.update(reportVal)
 
         return returnVal